Address code review feedback: tests and documentation
- Add rustdoc example to `fetch_pr_for_branch` showing async usage
- Refactor `GitRepoFixture` in ref_parser.rs to not mutate global cwd:
  - Remove `set_current_dir` calls from fixture constructors
  - Add `path()` method to access temp directory
  - Add test helpers (`current_branch_in`, `repo_from_fetch_head_in`,
    `repo_from_origin_in`) that operate on specific directories
  - Remove `#[serial]` from tests that no longer need it
- Fix git init compatibility: use `-c init.defaultBranch=main` pattern
  consistently instead of `--initial-branch=main`
- Refactor tests/e2e.rs into module directory structure:
  - Create tests/e2e/ with main.rs, common.rs, pr_basic.rs, pr_auto_detect.rs
  - Move shared infrastructure to common.rs
  - Group tests by functionality (basic PR vs auto-detect)
- Fix unused `git_repo_with_fetch_head` fixture by using it via rstest
- Add PR number assertion to fork disambiguation test to verify correct
  PR (#200) is selected based on head repository owner
- Add explanatory comment for `#[allow(dead_code)]` in tests/utils/mod.rs
  explaining why `#[expect]` cannot be used (multiple test crates compile
  the same utils module independently)
