+++ tests/e2e/main.rs
@@ -1,0 +1,5 @@
+//! End-to-end tests for vk PR command functionality.
+
+mod common;
+mod pr_auto_detect;
+mod pr_basic;
